+++ cs_form.js
@@ -3,7 +3,10 @@ export default {
   'form.validator.mustBeNumber': 'Musí být číslo',
   'form.validator.minNumber': 'Mělo by být větší než {min}',
   'form.validator.maxNumber': 'Mělo by být menší než {max}',
-  'form.validator.maxLength': 'Mělo by mít méně než {max} písmen',
+  'form.validator.maxLength': `Mělo by mít méně než {max} {what, select,
+    characters {písmen}
+    tags {tagů}
+  }`,
   'form.validator.minLength': 'Mělo by mít více než {min} písmen',
   'form.validator.email': 'Musí být validní e-mail',
   'form.validator.url': 'Musí být validní URL',