+++ en_form.js
@@ -3,7 +3,10 @@ export default {
   'form.validator.mustBeNumber': 'Must be a number',
   'form.validator.minNumber': 'Should be greater than {min}',
   'form.validator.maxNumber': 'Should be lower than {max}',
-  'form.validator.maxLength': 'Should be less than {max} characters',
+  'form.validator.maxLength': `Should be less than {max} {what, select,
+    characters {characters}
+    tags {tags}
+  }`,
   'form.validator.minLength': 'Should be greater than {min} characters',
   'form.validator.email': 'Must be a valid e-mail',
   'form.validator.url': 'Must be a valid URL',